Scoping bids, contracts, and offer_contract obj
Uses oslo_context to scope queries to the bids,
contracts, and offer_contract relationship objects. Updated and added
new tests to accomodate these changes. TG-163

+++ flocx_market/db/sqlalchemy/api.py
@@ -44,15 +44,13 @@ def offer_get(marketplace_offer_id, context):
             marketplace_offer_id=marketplace_offer_id).first()
 
 
-def offer_get_all_by_project_id(context):
+def offer_get_all(context):
+    return get_session().query(models.Offer).all()
 
+
+def offer_get_all_by_project_id(context):
     return get_session().query(models.Offer).filter_by(
         project_id=context.project_id).all()
-
-
-def offer_get_all(context):
-
-    return get_session().query(models.Offer).all()
 
 
 def offer_get_all_unexpired(context):
@@ -78,7 +76,8 @@ def offer_create(values, context):
 def offer_update(marketplace_offer_id, values, context):
 
     if context.is_admin:
-        offer_ref = offer_get(marketplace_offer_id, context)
+        offer_ref = get_session().query(models.Offer).filter_by(
+                    marketplace_offer_id=marketplace_offer_id).first()
     else:
         offer_ref = get_session().query(models.Offer).filter_by(
             marketplace_offer_id=marketplace_offer_id,
@@ -95,7 +94,8 @@ def offer_update(marketplace_offer_id, values, context):
 
 def offer_destroy(marketplace_offer_id, context):
     if context.is_admin:
-        offer_ref = offer_get(marketplace_offer_id, context)
+        offer_ref = get_session().query(models.Offer).filter_by(
+                    marketplace_offer_id=marketplace_offer_id).first()
     else:
         offer_ref = get_session().query(models.Offer).filter_by(
             marketplace_offer_id=marketplace_offer_id,
@@ -110,6 +110,8 @@ def offer_destroy(marketplace_offer_id, context):
             get_session().query(models.Offer).filter_by(
                 marketplace_offer_id=marketplace_offer_id,
                 project_id=context.project_id).delete()
+    else:
+        return None
 
 
 def bid_get(marketplace_bid_id, context):
@@ -117,17 +119,23 @@ def bid_get(marketplace_bid_id, context):
         marketplace_bid_id=marketplace_bid_id).first()
 
 
+def bid_get_all(context):
+    return get_session().query(models.Bid).all()
+
+
+def bid_get_all_by_project_id(context):
+    return get_session().query(models.Bid)\
+        .filter_by(project_id=context.project_id).all()
+
+
 def bid_get_all_unexpired(context):
     return get_session().query(models.Bid)\
         .filter(models.Bid.status != 'expired').all()
 
 
-def bid_get_all(context):
-    return get_session().query(models.Bid).all()
-
-
 def bid_create(values, context):
     values['marketplace_bid_id'] = uuidutils.generate_uuid()
+    values['project_id'] = context.project_id
     bid_ref = models.Bid()
     bid_ref.update(values)
     bid_ref.save(get_session())
@@ -135,18 +143,43 @@ def bid_create(values, context):
 
 
 def bid_update(marketplace_bid_id, values, context):
-    bid_ref = bid_get(marketplace_bid_id, context)
-    values.pop('marketplace_bid_id', None)
-    bid_ref.update(values)
-    bid_ref.save(get_session())
-    return bid_ref
+    if context.is_admin:
+        bid_ref = get_session().query(models.Bid).filter_by(
+                    marketplace_bid_id=marketplace_bid_id).first()
+    else:
+        bid_ref = get_session().query(models.Bid).filter_by(
+            marketplace_bid_id=marketplace_bid_id,
+            project_id=context.project_id).first()
+
+    if bid_ref:
+        values.pop('marketplace_bid_id', None)
+        bid_ref.update(values)
+        bid_ref.save(get_session())
+        return bid_ref
+    else:
+        return None
 
 
 def bid_destroy(marketplace_bid_id, context):
-    bid_ref = bid_get(marketplace_bid_id, context)
+    if context.is_admin:
+        bid_ref = get_session().query(models.Bid).filter_by(
+                    marketplace_bid_id=marketplace_bid_id).first()
+    else:
+        bid_ref = get_session().query(models.Bid).filter_by(
+            marketplace_bid_id=marketplace_bid_id,
+            project_id=context.project_id).first()
+
     if bid_ref:
-        get_session().query(models.Bid).filter_by(
-            marketplace_bid_id=marketplace_bid_id).delete()
+
+        if context.is_admin:
+            get_session().query(models.Bid).filter_by(
+                marketplace_bid_id=marketplace_bid_id).delete()
+        else:
+            get_session().query(models.Bid).filter_by(
+                marketplace_bid_id=marketplace_bid_id,
+                project_id=context.project_id).delete()
+    else:
+        return None
 
 
 # contract
@@ -165,42 +198,60 @@ def contract_get_all_unexpired(context):
 
 
 def contract_create(values, context):
-    values['contract_id'] = uuidutils.generate_uuid()
-    # exception for foreign key constraint needed here
-    offers = values['offers']
 
-    del values['offers']
-    contract_ref = models.Contract()
-    contract_ref.update(values)
-    contract_ref.save(get_session())
+    if context.is_admin:
+        values['contract_id'] = uuidutils.generate_uuid()
+        # exception for foreign key constraint needed here
+        offers = values['offers']
 
-    # update foreign key for offers
-    for offer_id in offers:
-        ocr_data = dict(contract_id=values['contract_id'],
-                        marketplace_offer_id=offer_id,
-                        status='unretrieved')
-        offer_contract_relationship_create(ocr_data)
-    return contract_ref
+        del values['offers']
+        contract_ref = models.Contract()
+        contract_ref.update(values)
+        contract_ref.save(get_session())
+
+        # update foreign key for offers
+        for offer_id in offers:
+            ocr_data = dict(contract_id=values['contract_id'],
+                            marketplace_offer_id=offer_id,
+                            status='unretrieved')
+            offer_contract_relationship_create(context, ocr_data)
+        return contract_ref
+    else:
+        return None
 
 
 def contract_update(contract_id, values, context):
-    contract_ref = contract_get(contract_id, context)
-    values.pop('contract_id', None)
-    contract_ref.update(values)
-    contract_ref.save(get_session())
-    return contract_ref
+
+    if context.is_admin:
+        contract_ref = get_session().query(models.Contract).filter_by(
+                        contract_id=contract_id).first()
+        if contract_ref:
+            values.pop('contract_id', None)
+            contract_ref.update(values)
+            contract_ref.save(get_session())
+            return contract_ref
+        else:
+            return None
+    else:
+        return None
 
 
 def contract_destroy(contract_id, context):
-    contract_ref = contract_get(contract_id, context)
-    if contract_ref:
-        get_session().query(models.Contract).filter_by(
-            contract_id=contract_id).delete()
+    if context.is_admin:
+        contract_ref = contract_get(contract_id, context)
+        if contract_ref:
+            get_session().query(models.Contract).filter_by(
+                contract_id=contract_id).delete()
+        else:
+            return None
+    else:
+        return None
 
 
 # offer_contract_relationship
-def offer_contract_relationship_get(marketplace_offer_id=None,
-                                    contract_id=None):
+def offer_contract_relationship_get(context,
+                                    contract_id=None,
+                                    marketplace_offer_id=None):
 
     if (contract_id is not None) and (marketplace_offer_id is None):
         return get_session().query(models.OfferContractRelationship).filter_by(
@@ -212,49 +263,68 @@ def offer_contract_relationship_get(marketplace_offer_id=None,
     elif (contract_id is not None) and (marketplace_offer_id is not None):
         return get_session().query(models.OfferContractRelationship) \
             .filter(models.OfferContractRelationship.contract_id
-                    == contract_id and
+                    == contract_id,
                     models.OfferContractRelationship.marketplace_offer_id
                     == marketplace_offer_id).first()
 
 
-def offer_contract_relationship_get_all():
+def offer_contract_relationship_get_all(context):
     return get_session().query(models.OfferContractRelationship).all()
 
 
-def offer_contract_relationship_get_all_unexpired():
+def offer_contract_relationship_get_all_unexpired(context):
     return get_session().query(models.OfferContractRelationship)\
         .filter(models.OfferContractRelationship.status != 'expired').all()
 
 
-def offer_contract_relationship_create(values):
-    values['offer_contract_relationship_id'] = uuidutils.generate_uuid()
-    # exception for foreign key constraint needed here
-    offer_contract_relationship_ref = models.OfferContractRelationship()
-    offer_contract_relationship_ref.update(values)
-    offer_contract_relationship_ref.save(get_session())
-    return offer_contract_relationship_ref
+def offer_contract_relationship_create(context, values):
+
+    if context.is_admin:
+        values['offer_contract_relationship_id'] = uuidutils.generate_uuid()
+        # exception for foreign key constraint needed here
+        offer_contract_relationship_ref = models.OfferContractRelationship()
+        offer_contract_relationship_ref.update(values)
+        offer_contract_relationship_ref.save(get_session())
+        return offer_contract_relationship_ref
+    else:
+        return None
 
 
-def offer_contract_relationship_update(contract_id,
-                                       marketplace_offer_id, values):
-    offer_contract_relationship_ref = offer_contract_relationship_get(
-        contract_id, marketplace_offer_id)
+def offer_contract_relationship_update(context,
+                                       contract_id,
+                                       marketplace_offer_id,
+                                       values):
+    if context.is_admin:
+        offer_contract_relationship_ref = offer_contract_relationship_get(
+            context,
+            contract_id=contract_id,
+            marketplace_offer_id=marketplace_offer_id)
 
-    values.pop('contract_id', None)
-    values.pop('marketplace_offer_id', None)
-    offer_contract_relationship_ref.update(values)
-    offer_contract_relationship_ref.save(get_session())
-    return offer_contract_relationship_ref
+        values.pop('contract_id', None)
+        values.pop('marketplace_offer_id', None)
+        offer_contract_relationship_ref.update(values)
+        offer_contract_relationship_ref.save(get_session())
+        return offer_contract_relationship_ref
+    else:
+        return None
 
 
-def offer_contract_relationship_destroy(contract_id,
+def offer_contract_relationship_destroy(context,
+                                        contract_id,
                                         marketplace_offer_id):
-    offer_contract_relationship_ref = offer_contract_relationship_get(
-        contract_id, marketplace_offer_id)
+    if context.is_admin:
+        offer_contract_relationship_ref = offer_contract_relationship_get(
+                                            context,
+                                            contract_id,
+                                            marketplace_offer_id)
 
-    if offer_contract_relationship_ref:
-        get_session().query(models.OfferContractRelationship) \
-            .filter(models.OfferContractRelationship.contract_id
-                    == contract_id and
-                    models.OfferContractRelationship.marketplace_offer_id
-                    == marketplace_offer_id).delete()
+        if offer_contract_relationship_ref:
+            get_session().query(models.OfferContractRelationship) \
+                .filter(models.OfferContractRelationship.contract_id
+                        == contract_id and
+                        models.OfferContractRelationship.marketplace_offer_id
+                        == marketplace_offer_id).delete()
+        else:
+            return None
+    else:
+        return None
